Delete dead OpenAPI catalog path
